Move register_canonicalize to graph.rewriting.utils
